Implementation of derived data logic
This commit includes the logic for derived data calculation. The logic is simple and can be improved on the future, but take in consideration the case that there is no sample on the pass and regular ones. Also what to do on corner cases, as for example the addition of new stations