Split on multi microsystems solution and Spring dependency injection fixed
Till now had been in troubles for doing correctly the autowired. Also had been scheduled on a main program that will be on charge of spreading each process on different threads.

On this version it is throwing exceptions due to the connectivity with the database. And will be fixed on the next commit