Makes GisDatabaseService implementation stateless
With this commit we will allow to use Spring autowired on the GisDatabaseService. This is fundamental as we are using the jpa repositories. It was working until for the other two databaservice as the respotiory was done without JPA.

We are going to keep working with the template pattern but in this case we are going to pass the state on the function parameter. We will skip to have internal class variables and also to set that values.

On the other hand we are executing twice the locationcall to the customer microservice (needs to improve)