GisDatabaseService for Raw
On this commit we deliver the feature for inserting updating data on the Gis - StationSystemRaw table. By now we develop the main logic but not implemented yet the tables.

As it has been developed the logical level for inserting or updating it is still missing the hibernate layer, but alhambdulila, piece of cake is missing