Extend all CSV files from a Base class that contains times
In order to disacople the CSV objects we are creating specific CSV files for each ones of the outputs.

It means as well, that we will need to alter the Service class in order to adapt to this new POJO.

What CSV pojos does is extend from BaseCsv and have as a parameter of the constructor the data class, this will be mapped to the CSV columns and also call to the default constructor of the BaseCsv, that will set the time.